Initial commit
Added a README, which outlines project goals and the like. The readme is
written in past-tense, because I plan on continuing my d3 learning way
into the future.

I'll state for the record that I really want this project to be a
standard level of best-practice when it comes to code readability and
repository structure.