Started main program and added the curvetypes row
First off, we grab data from the window (defined earlier) and the main
function is fired when the DOM loads.

Next, we grab each row of controls (for dataset, curvetype and settings)
and start rendering them.

So far, we've gotten as far as rendering the curvetype row - but
parameter configuration is still missing (I need a slider).
Also, I chose to add 'select all' and 'select none' buttons for
convenience.