Added settings row rendering and styles
Continuing with rendering each control row, but this time settings are
being rendered. I have used a subscriber pattern to monitor what the
settings are as they change - perhaps I'll use this pattern for every
row, if it fits the implementation.

Also did some simple styling on the settings controls