Changed curvetype row to subscriber pattern
Now, we have on-time updates for when the ui changes. Each argument is
now properly rendered, although we don't do anything with them yet. So
far, the selection buttons work as expected and the labels are all
correct.

Each argument now has a default value associated with it, taken from the
documentation for d3.

As a side note, additional styling was done for the arguments, including
a global modification on input[type=number] styles