Defaults and Rendering Start
Now each of curvetypes, settings and datasets returns an object with a
subscriber function and a default value. This will be useful for initial
renders.

Also started work on the function to render / animate the data. The
render function will have a few lower-level functions - I have added two
so far: generateData and getExtentFromPoints.