Updated the easing to be more playful
Now, instead of boring easing functions, we have bouncy and playful
stuff that also looks really cool. Also changed the timing of the
point-easeing-in bit so that it happens sequentially, but (no matter how
many points) will always take a set amount of time.

Note that, although verbose, the timing math is easiest to understand
when written like it is.