Styling and Animation Enhancements
The points come in from the center of the graph now, and each curvetype
is colored differently, depending on its arguments (or randomly if it
has none). Also, animation is now smoother because fadeins and outs
happen at the same time movement occurs.

Side note, I added some methods to the Chart class for handling data and
coloring.